Fixed warning issue

+++ Sources/Enemy/Enemy.cpp
@@ -964,7 +964,7 @@ void Enemy::LoadWeapon(bool left, std::string weaponFile)
 			if (m_pVoxelCharacter->IsLeftWeaponLoaded())
 			{
 				// Lights
-				for (int i = 0; i < m_pVoxelCharacter->GetLeftWeapon()->GetNumLights(); ++i)
+				for (size_t i = 0; i < m_pVoxelCharacter->GetLeftWeapon()->GetNumLights(); ++i)
 				{
 					unsigned int lightID;
 					glm::vec3 lightPos;
@@ -983,7 +983,7 @@ void Enemy::LoadWeapon(bool left, std::string weaponFile)
 				}
 
 				// Particle effects
-				for (int i = 0; i < m_pVoxelCharacter->GetLeftWeapon()->GetNumParticleEffects(); ++i)
+				for (size_t i = 0; i < m_pVoxelCharacter->GetLeftWeapon()->GetNumParticleEffects(); ++i)
 				{
 					unsigned int particleEffectID;
 					glm::vec3 particleEffectPos;
@@ -1011,7 +1011,7 @@ void Enemy::LoadWeapon(bool left, std::string weaponFile)
 			if (m_pVoxelCharacter->IsRightWeaponLoaded())
 			{
 				// Lights
-				for (int i = 0; i < m_pVoxelCharacter->GetRightWeapon()->GetNumLights(); ++i)
+				for (size_t i = 0; i < m_pVoxelCharacter->GetRightWeapon()->GetNumLights(); ++i)
 				{
 					unsigned int lightID;
 					glm::vec3 lightPos;
@@ -1030,7 +1030,7 @@ void Enemy::LoadWeapon(bool left, std::string weaponFile)
 				}
 
 				// Particle effects
-				for (int i = 0; i < m_pVoxelCharacter->GetRightWeapon()->GetNumParticleEffects(); ++i)
+				for (size_t i = 0; i < m_pVoxelCharacter->GetRightWeapon()->GetNumParticleEffects(); ++i)
 				{
 					unsigned int particleEffectID;
 					glm::vec3 particleEffectPos;
@@ -1083,7 +1083,7 @@ void Enemy::UnloadWeapon(bool left) const
 		if (isWeaponLoaded)
 		{
 			// Lights
-			for (int i = 0; i < pWeapon->GetNumLights(); ++i)
+			for (size_t i = 0; i < pWeapon->GetNumLights(); ++i)
 			{
 				unsigned int lightID;
 				glm::vec3 lightPos;
@@ -1123,7 +1123,7 @@ void Enemy::UnloadWeapon(bool left) const
 			}
 
 			// Particle Effects
-			for (int i = 0; i < pWeapon->GetNumParticleEffects(); ++i)
+			for (size_t i = 0; i < pWeapon->GetNumParticleEffects(); ++i)
 			{
 				unsigned int particleEffectID;
 				glm::vec3 particleEffectPos;
@@ -2862,7 +2862,7 @@ void Enemy::UpdateWeaponLights() const
 		{
 			if (isWeaponLoaded)
 			{
-				for (int j = 0; j < pWeapon->GetNumLights(); ++j)
+				for (size_t j = 0; j < pWeapon->GetNumLights(); ++j)
 				{
 					unsigned int lightID;
 					glm::vec3 lightPos;
@@ -2929,7 +2929,7 @@ void Enemy::UpdateWeaponParticleEffects() const
 		{
 			if (isWeaponLoaded)
 			{
-				for (int j = 0; j < pWeapon->GetNumParticleEffects(); ++j)
+				for (size_t j = 0; j < pWeapon->GetNumParticleEffects(); ++j)
 				{
 					unsigned int particleEffectID;
 					glm::vec3 particleEffectPos;
